clean up renaming relics
