+++ app-android/src/main/res/values/strings.xml
@@ -1,3 +1,4 @@
 <resources>
-    <string name="app_name">MastodonCompose</string>
+    <string name="app_name">Dodo</string>
+    <string name="app_name_fullname">Dodo for Mastodon</string>
 </resources>
